requirements.txt
# Optimization Core
pulp>=2.7.0
numpy>=1.24.0
scipy>=1.10.0
gurobipy>=10.0.0 # Optional, falls back to CBC in PuLP

# Data & Networking
websockets>=11.0
httpx>=0.24.0
curl_cffi>=0.5.0
pandas>=2.0.0
influxdb-client>=1.36.0

# Protocol / Web3
eth-account>=0.8.0
web3>=6.0.0

# Polymarket SDK
py-clob-client>=0.18.0

# AI / LLM Integration
openai>=1.0.0
chromadb>=0.4.0  # Optional: for semantic caching
sentence-transformers>=2.2.0  # Optional: for local embeddings (Hacha Protocol)

# Utils
python-telegram-bot>=20.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
rich>=13.0.0
rapidfuzz>=3.0.0
